feat: multiple instances of the server can be started
